+++ src/controller/timeline.rs
@@ -3,6 +3,7 @@ use std::collections::VecDeque;
 use crate::app_context::AppContext;
 use crate::session;
 use crate::usecase;
+use crate::usecase::message::Post;
 use crate::{views, views::TemplateToResponse};
 use actix_session::Session;
 use actix_web::{web, HttpResponse, Responder};
@@ -31,7 +32,7 @@ pub async fn timeline(context: web::Data<AppContext>, session: Session) -> impl 
         Err(e) => println!("{:#?}", e),
     };
 
-    let mut posts: VecDeque<views::timeline::Post> = VecDeque::new();
+    let mut posts: VecDeque<Post> = VecDeque::new();
 
     for one_message in messages {
         let message_id = match one_message.try_get::<String>("", "message_id") {
@@ -58,7 +59,7 @@ pub async fn timeline(context: web::Data<AppContext>, session: Session) -> impl 
             Ok(res) => res,
             Err(e) => return HttpResponse::InternalServerError().body(e.to_string()),
         };
-        posts.push_front(views::timeline::Post {
+        posts.push_front(Post {
             message_id,
             user_name,
             message_text,
